+++ packages/sls-env/src/__tests__/dependencies.test.ts
@@ -1,13 +1,5 @@
-import { APIGatewayProxyEventV2 } from 'aws-lambda'
 import { EventDependencyError } from '../error/index'
-import {
-  APIGatewayEventDependency,
-  EventDependency,
-  Handler,
-  apiGatewayEventDependency,
-  environment,
-  eventDependency
-} from '..'
+import { EventDependency, Handler, environment, eventDependency } from '..'
 
 describe('serverless environment', () => {
   type MessageEvent = { message: string }
@@ -181,40 +173,6 @@ describe('serverless environment', () => {
       await expect(handler({ message: 'client3' }, { name: 'world' })).rejects.toThrow(
         new EventDependencyError("No event based dependency found for 'buildMessage'")
       )
-    })
-
-    it('changes event based dependencies based on the payload - APIGateway event', async () => {
-      type BuildMessage = (message: string, name: string) => string
-      type BuildMessageDependencies = {
-        buildMessage: APIGatewayEventDependency<BuildMessage, EventPayload>
-      }
-
-      const buildMessages: Record<string, BuildMessage> = {
-        'test.com': (_m, _n) => `Using client for test.com!`
-      }
-
-      const event = ({
-        headers: {
-          host: 'test.com'
-        }
-      } as unknown) as APIGatewayProxyEventV2
-
-      const handler = environment<
-        Handler<APIGatewayProxyEventV2, NameContext, string>,
-        any,
-        BuildMessageDependencies,
-        EventPayload
-      >()
-        .global({
-          buildMessage: apiGatewayEventDependency(buildMessages, ({ event }) => event.headers['host'] as string)
-        })
-        .app(({ payload: { event, context }, dependencies: { buildMessage } }) =>
-          buildMessage(event.message, context.name)
-        )
-        .successHandler(({ result }) => result as string).start
-
-      const response1 = await handler(event, { name: 'world' })
-      expect(response1).toBe('Using client for test.com!')
     })
   })
 })
